Add files via upload
The year 2020 was very difficult for all of us and we hope that in 2021 things will get better. I created a PCB that was inspired by a snowflake with LED's all over its outline and in its center the message of Merry Christmas. The intention is to illuminate our Christmas and bring light for the coming year. Made for the contest held by https://jlcpcb.com/